platform/api/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
pydantic==2.10.3
pytest==8.3.4
httpx==0.28.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
neo4j==5.27.0
boto3==1.35.70
PyJWT==2.10.1
PyYAML==6.0.2
prometheus-client==0.21.1
alembic==1.14.1
